add op check method
