+++ runtime/onert/core/include/backend/ValidatorBase.h
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2025 Samsung Electronics Co., Ltd. All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef __ONERT_BACKEND_VALIDATOR_BASE_H__
+#define __ONERT_BACKEND_VALIDATOR_BASE_H__
+
+#include "ir/Graph.h"
+#include "ir/OperationVisitor.h"
+
+namespace onert::backend
+{
+
+class ValidatorBase : public ir::OperationVisitor
+{
+public:
+  virtual ~ValidatorBase() = default;
+  ValidatorBase() = delete;
+  ValidatorBase(const ir::Graph &graph) : _graph(graph), _supported(false) {}
+
+public:
+  virtual bool supported(const ir::IOperation &op) final
+  {
+    op.accept(*this);
+    return _supported;
+  }
+
+protected:
+  using OperationVisitor::visit;
+
+#define OP(InternalName) \
+  void visit(const ir::operation::InternalName &) override { _supported = false; }
+#include "ir/Operations.lst"
+#undef OP
+
+protected:
+  const ir::Graph &_graph;
+  bool _supported;
+};
+
+} // namespace onert::backend
+
+#endif // __ONERT_BACKEND_VALIDATOR_BASE_H__
